+++ geowombat/backends/xarray_.py
@@ -9,6 +9,7 @@ from .rasterio_ import get_ref_image_meta, warp, warp_images, get_file_bounds, w
 from .rasterio_ import transform_crs as rio_transform_crs
 
 import numpy as np
+from rasterio import open as rio_open
 from rasterio.windows import Window
 from rasterio.coords import BoundingBox
 import dask.array as da
@@ -213,6 +214,9 @@ def warp_open(filename,
 
     ref_kwargs = _check_config_globals(filename, 'reference', ref_kwargs)
 
+    with rio_open(filename) as src:
+        tags = src.tags()
+
     with xr.open_rasterio(warp(filename,
                                resampling=resampling,
                                **ref_kwargs),
@@ -236,8 +240,10 @@ def warp_open(filename,
                     # Avoid nested opens within a `config` context
                     if len(new_band_names) != len(src.band.values.tolist()):
 
-                        logger.warning('  The new bands, {}, do not match the sensor bands, {}.'.format(new_band_names,
-                                                                                                        src.band.values.tolist()))
+                        if not src.gw.config['ignore_warnings']:
+
+                            logger.warning('  The new bands, {}, do not match the sensor bands, {}.'.format(new_band_names,
+                                                                                                            src.band.values.tolist()))
 
                     else:
 
@@ -259,6 +265,11 @@ def warp_open(filename,
 
         src.attrs['filename'] = filename
         src.attrs['resampling'] = resampling
+
+        if tags:
+            attrs = src.attrs.copy()
+            attrs.update(tags)
+            src = src.assign_attrs(**attrs)
 
         if dtype:
 
@@ -327,6 +338,9 @@ def mosaic(filenames,
 
     footprints = []
 
+    with rio_open(filenames[0]) as src_:
+        tags = src_.tags()
+
     # Combine the data
     with xr.open_rasterio(warped_objects[0], **kwargs) as darray:
 
@@ -335,7 +349,6 @@ def mosaic(filenames,
         # Get the original bounds, unsampled
         with xr.open_rasterio(filenames[0], **kwargs) as src_:
             footprints.append(src_.gw.geometry)
-        src_ = None
 
         for fidx, fn in enumerate(warped_objects[1:]):
 
@@ -407,6 +420,11 @@ def mosaic(filenames,
         darray.attrs['filename'] = [Path(fn).name for fn in filenames]
         darray.attrs['resampling'] = resampling
 
+        if tags:
+            attrs = darray.attrs.copy()
+            attrs.update(tags)
+            darray = darray.assign_attrs(**attrs)
+
         darray.gw.footprint_grid = footprints
 
         if dtype:
@@ -475,12 +493,15 @@ def concat(filenames,
 
     ref_kwargs = _check_config_globals(filenames, bounds_by, ref_kwargs)
 
+    with rio_open(filenames[0]) as src_:
+        tags = src_.tags()
+
     # Keep a copy of the transformed attributes.
     with xr.open_rasterio(warp(filenames[0],
                                resampling=resampling,
-                               **ref_kwargs), **kwargs) as ds_:
+                               **ref_kwargs), **kwargs) as src_:
 
-        attrs = ds_.attrs.copy()
+        attrs = src_.attrs.copy()
 
     if time_names:
 
@@ -528,54 +549,62 @@ def concat(filenames,
         output = xr.concat(concat_list, dim=stack_dim.lower())
 
         # Assign the new time band names
-        ds = output.assign_coords(time=new_time_names)
+        src = output.assign_coords(time=new_time_names)
 
     else:
 
         # Warp all images and concatenate along
         #   the 'time' axis into a DataArray.
-        ds = xr.concat([xr.open_rasterio(warp(fn,
-                                              resampling=resampling,
-                                              **ref_kwargs), **kwargs)
-                        for fn in filenames], dim=stack_dim.lower())
+        src = xr.concat([xr.open_rasterio(warp(fn,
+                                               resampling=resampling,
+                                               **ref_kwargs), **kwargs)
+                         for fn in filenames], dim=stack_dim.lower())
 
-    ds = ds.assign_attrs(**attrs)
+    src.attrs['filename'] = [Path(fn).name for fn in filenames]
+
+    if tags:
+        attrs = src.attrs.copy()
+        attrs.update(tags)
+
+    src = src.assign_attrs(**attrs)
 
     if not time_names and (stack_dim == 'time'):
-        ds.coords['time'] = parse_filename_dates(filenames)
+        src.coords['time'] = parse_filename_dates(filenames)
 
     if band_names:
-        ds.coords['band'] = band_names
+        src.coords['band'] = band_names
     else:
 
-        if ds.gw.sensor:
+        if src.gw.sensor:
 
-            if ds.gw.sensor not in ds.gw.avail_sensors:
+            if src.gw.sensor not in src.gw.avail_sensors:
 
-                logger.warning('  The {} sensor is not currently supported.\nChoose from [{}].'.format(ds.gw.sensor,
-                                                                                                       ', '.join(ds.gw.avail_sensors)))
+                logger.warning('  The {} sensor is not currently supported.\nChoose from [{}].'.format(src.gw.sensor,
+                                                                                                       ', '.join(src.gw.avail_sensors)))
 
             else:
 
-                new_band_names = list(ds.gw.wavelengths[ds.gw.sensor]._fields)
+                new_band_names = list(src.gw.wavelengths[src.gw.sensor]._fields)
 
-                if len(new_band_names) != len(ds.band.values.tolist()):
+                if len(new_band_names) != len(src.band.values.tolist()):
 
-                    logger.warning('  The new bands, {}, do not match the sensor bands, {}.'.format(new_band_names,
-                                                                                                    ds.band.values.tolist()))
+                    if not src.gw.config['ignore_warnings']:
+
+                        logger.warning('  The new bands, {}, do not match the sensor bands, {}.'.format(new_band_names,
+                                                                                                        src.band.values.tolist()))
 
                 else:
 
-                    ds.coords['band'] = new_band_names
-                    ds.attrs['sensor'] = ds.gw.sensor_names[ds.gw.sensor]
+                    src.coords['band'] = new_band_names
+                    src.attrs['sensor'] = src.gw.sensor_names[src.gw.sensor]
 
     if dtype:
         
-        attrs = ds.attrs.copy()
-        return ds.astype(dtype).assign_attrs(**attrs)
+        attrs = src.attrs.copy()
+        return src.astype(dtype).assign_attrs(**attrs)
 
     else:
-        return ds
+        return src
 
 
 def transform_crs(data_src,