+++ geowombat/tasks/pipeline.py
@@ -9,7 +9,8 @@ from datetime import datetime
 # from inspect import signature
 
 from ..errors import logger
-import geowombat as gw
+from .. import config as gw_config
+from .. import open as gw_open
 
 import xarray as xr
 import graphviz
@@ -63,7 +64,7 @@ class BaseGeoTask(ABC):
                  inputs,
                  outputs,
                  tasks,
-                 clean,
+                 clean=None,
                  config_args=None,
                  open_args=None,
                  func_args=None,
@@ -73,7 +74,7 @@ class BaseGeoTask(ABC):
         self.inputs = inputs
         self.outputs = outputs
         self.tasks = tasks
-        self.clean = clean
+        self.clean = clean if clean else {}
         self.config_args = config_args if inputs else {}
         self.open_args = open_args if inputs else {}
         self.func_args = func_args if inputs else {}
@@ -131,7 +132,7 @@ class BaseGeoTask(ABC):
                        out_args=self_out_args_copy)
 
     @abstractmethod
-    def execute(self, task_id, task, src, **kwargs):
+    def execute(self, task_id, task, src, task_results, attrs, **kwargs):
         """Execute a task"""
         pass
 
@@ -166,7 +167,7 @@ class BaseGeoTask(ABC):
         task_output = self.outputs[task_id]
         when = datetime.now().strftime('%A, %d-%m-%Y at %H:%M:%S')
 
-        task_log = f'{when} | {task_output} | {task_id-random_id}'
+        task_log = f"{when} | {task_output} | task_id-{random_id}"
 
         return f'{task_log} ok\n' if self._check_task(task_id) else f'{task_log} failed\n'
 
@@ -178,7 +179,11 @@ class BaseGeoTask(ABC):
                 lines = f.readlines()
 
         else:
+
             lines = []
+
+            with open(self.log_file, mode='w') as f:
+                f.writelines(lines)
 
         with open(self.log_file, mode='r+') as f:
 
@@ -196,7 +201,10 @@ class GraphBuilder(object):
         https://github.com/benbovy/xarray-simlab/blob/master/xsimlab/dot.py
     """
 
-    def visualize(self):
+    def visualize(self, **kwargs):
+
+        if not kwargs:
+            kwargs = {'rankdir': 'LR'}
 
         self.seen = set()
         self.inputs_seen = set()
@@ -205,7 +213,7 @@ class GraphBuilder(object):
         counter = 0
 
         self.g = graphviz.Digraph()
-        self.g.subgraph(graph_attr={'rankdir': 'LR'})
+        self.g.subgraph(graph_attr=kwargs)
 
         for task_id, task in self.tasks:
 
@@ -215,7 +223,20 @@ class GraphBuilder(object):
                 self.g.node(task_id, label=f'Task {task_id}: {task.__name__}', **PROC_NODE_ATTRS)
 
             if task_id != list(self.tasks)[0][0]:
-                self.g.edge(list(self.tasks)[counter - 1][0], task_id, **PROC_EDGE_ATTRS)
+
+                if isinstance(self.inputs[task_id], str):
+                    self.g.edge(list(self.tasks)[counter-1][0], task_id, **PROC_EDGE_ATTRS)
+                else:
+
+                    task_list_ = list(list(zip(*self.tasks))[0])
+
+                    for ctask in self.inputs[task_id]:
+                        if ctask in task_list_:
+                            cidx = task_list_.index(ctask)
+                        else:
+                            cidx = counter-1
+
+                        self.g.edge(list(self.tasks)[cidx][0], task_id, **PROC_EDGE_ATTRS)
 
             for config_key, config_setting in self.config_args.items():
 
@@ -257,18 +278,22 @@ class GraphBuilder(object):
 
             for out_key, out_setting in self.out_args.items():
 
-                with self.g.subgraph(name='cluster_1') as c:
+                if not output_.startswith('mem|'):
 
-                    c.attr(style='filled', color='#a6d5ab')
-                    c.node(out_key, label=f'{out_key}: {out_setting}', **OUT_NODE_ATTRS)
-                    c.attr(label='geowombat.to_raster() args')
+                    with self.g.subgraph(name='cluster_1') as c:
 
-                self.g.edge(out_key, f'{task_id} {self.outputs[task_id]}', **OUT_EDGE_ATTRS)
+                        c.attr(style='filled', color='#a6d5ab')
+                        c.node(out_key, label=f'{out_key}: {out_setting}', **OUT_NODE_ATTRS)
+                        c.attr(label='geowombat.to_raster() args')
+
+                    self.g.edge(out_key, f'{task_id} {self.outputs[task_id]}', **OUT_EDGE_ATTRS)
 
             if counter > 0:
 
                 task_id_ = list(self.outputs.keys())[counter-1]
-                self.g.edge(f'{task_id_} {self.outputs[task_id_]}', task_id, weight='200', **edge_attrs)
+
+                if not self.outputs[task_id_].startswith('mem|'):
+                    self.g.edge(f'{task_id_} {self.outputs[task_id_]}', task_id, weight='200', **edge_attrs)
 
             counter += 1
 
@@ -296,32 +321,41 @@ class GraphBuilder(object):
 
         for input_ in input_list:
 
-            if input_ in self.outputs_seen:
-                task_id_b = None
+            if isinstance(input_, tuple) or isinstance(input_, list):
+                gen = input_
             else:
+                gen = (input_,)
 
-                if input_ not in self.inputs_seen:
+            for gen_item in gen:
 
-                    self.inputs_seen.add(input_)
-
-                    if input_ not in self.outputs:
-                        self.g.node(f'{task_id} {input_}', label=input_, **INPUT_NODE_ATTRS)
-
-                    task_id_b = task_id
-
+                if gen_item in self.outputs_seen:
+                    task_id_b = None
                 else:
-                    task_id_b = list(self.tasks)[counter-1][0]
 
-            if task_id_b:
+                    if gen_item not in self.inputs_seen:
 
-                if input_ not in self.outputs:
+                        self.inputs_seen.add(gen_item)
 
-                    for itask, iinputs in self.inputs.items():
-                        if input_ in iinputs:
-                            task_id_b = itask
-                            break
+                        gen_label = Path(gen_item).name if Path(gen_item).is_file() else gen_item
 
-                    self.g.edge(f'{task_id_b} {input_}', task_id, weight='200', **INPUT_EDGE_ATTRS)
+                        if gen_item not in self.outputs:
+                            self.g.node(f'{task_id} {gen_item}', label=gen_label, **INPUT_NODE_ATTRS)
+
+                        task_id_b = task_id
+
+                    else:
+                        task_id_b = list(self.tasks)[counter-1][0]
+
+                if task_id_b:
+
+                    if gen_item not in self.outputs:
+
+                        for itask, iinputs in self.inputs.items():
+                            if gen_item in iinputs:
+                                task_id_b = itask
+                                break
+
+                        self.g.edge(f'{task_id_b} {gen_item}', task_id, weight='200', **INPUT_EDGE_ATTRS)
 
 
 class GeoTask(BaseGeoTask, GraphBuilder):
@@ -342,34 +376,42 @@ class GeoTask(BaseGeoTask, GraphBuilder):
 
     Example:
         >>> import geowombat as gw
+        >>> from geowombat.data import l8_224078_20200518, l8_224078_20200518_B3, l8_224078_20200518_B4
         >>> from geowombat.tasks import GeoTask
         >>>
-        >>> from geowombat.radiometry import RadTransforms
-        >>> rt = RadTransforms()
+        >>> # Tasks a and take an image as input
+        >>> # Task c takes two images as input
+        >>> # Task d takes the result of tasks a, b, and c
+        >>> inputs = {'a': l8_224078_20200518, 'b': l8_224078_20200518, 'c': (l8_224078_20200518_B3, l8_224078_20200518_B4), 'd': ('a', 'b', 'c')}
         >>>
-        >>> tasks = (('A', rt.dn_to_sr), ('B', gw.ndvi))
-        >>> clean = {'A': 'task'}
+        >>> # The output task names
+        >>> # Tasks a, b, c, and d generate in-memory arrays
+        >>> outputs = {'a': 'mem|r1', 'b': 'mem|r2', 'c': 'mem|r3', 'd': 'mem|stack'}
         >>>
-        >>> inputs = {'A': ('input.tif', 'sza.tif', 'saa.tif', 'vza.tif', 'vaa.tif'),
-        >>>           'B': 'A'}
+        >>> # Task a and b execute the `norm_diff`
+        >>> # Task c computes the mean of c
+        >>> # Task d concatenates a, b, and c
+        >>> tasks = (('a', gw.norm_diff), ('b', gw.norm_diff), ('c', xr.DataArray.mean), ('d', xr.concat))
         >>>
-        >>> outputs = {'A': 'sr.tif',
-        >>>            'B': 'ndvi.tif'}
+        >>> # No intermediate outputs to cleanup
+        >>> clean = {}
         >>>
-        >>> func_args = {'A': {'meta': 'meta.mtl'}}
+        >>> # Task a and b require the bands to ratio
+        >>> # Task c and d require the dimension to reduce
+        >>> func_args = {'a': {'b1': 'green', 'b2': 'red'}, 'b': {'b1': 'blue', 'b2': 'green'}, 'c': {'dim': 'band'}, 'd': {'dim': 'band'}}
         >>>
         >>> open_args = {'chunks': 512}
-        >>> config_args = {'sensor': 'l7', 'scale_factor': 0.0001}
+        >>> config_args = {'sensor': 'bgr', 'nodata': 0, 'scale_factor': 0.0001}
         >>> out_args = {'compress': 'lzw', 'overwrite': True}
         >>>
-        >>> task = GeoTask(inputs, outputs, tasks, clean, config_args, open_args, func_args, out_args)
+        >>> # Setup the task
+        >>> task_mean = GeoTask(inputs, outputs, tasks, clean=clean, config_args=config_args, open_args=open_args, func_args=func_args, out_args=out_args)
         >>>
-        >>> task.visualize()
-        >>> task.submit()
+        >>> # Visualize the task
+        >>> task_mean.visualize()
         >>>
-        >>> # Add pipelines
-        >>> task_sum = GeoTask(...)
-        >>> task = task + task_sum
+        >>> # Submit the task
+        >>> task_mean.submit()
     """
 
     def __init__(self,
@@ -386,34 +428,44 @@ class GeoTask(BaseGeoTask, GraphBuilder):
         super().__init__(inputs,
                          outputs,
                          tasks,
-                         clean,
+                         clean=clean,
                          config_args=config_args,
                          open_args=open_args,
                          func_args=func_args,
                          out_args=out_args,
                          log_file=log_file)
 
-    def execute(self, task_id, task, src, **kwargs):
+    def execute(self, task_id, task, src, task_results, attrs, **kwargs):
 
         """
         Executes an individual task
 
         Args:
             task_id (str)
-            task (object)
+            task (func)
             src (DataArray | list)
+            task_results (dict)
+            attrs (dict)
             kwargs (Optional[dict])
         """
 
         # Execute the task
-        if isinstance(src, list):
-            res = task(*src, **kwargs)
+        if isinstance(src, tuple):
+            res = task((task_results[i] for i in src), **kwargs)
         else:
             res = task(src, **kwargs)
 
+        if not hasattr(res, 'band'):
+            res = res.expand_dims(dim='band').assign_coords({'band': ['res']})
+
         # Write to file
         if task_id in self.outputs:
-            res.gw.to_raster(self.outputs[task_id], **self.out_args)
+            if self.outputs[task_id].lower().endswith('.tif'):
+                if not hasattr(res, 'crs'):
+                    res.attrs = attrs
+                res.gw.to_raster(self.outputs[task_id], **self.out_args)
+
+        return res
 
     def submit(self):
 
@@ -421,7 +473,11 @@ class GeoTask(BaseGeoTask, GraphBuilder):
         Submits a pipeline task
         """
 
-        with gw.config.update(**self.config_args):
+        task_results = {}
+        attrs = None
+        res = None
+
+        with gw_config.update(**self.config_args):
 
             counter = 0
 
@@ -431,39 +487,45 @@ class GeoTask(BaseGeoTask, GraphBuilder):
                 kwargs = self.func_args[task_id] if task_id in self.func_args else {}
 
                 # Check task input(s)
-                if isinstance(self.inputs[task_id], str) and not Path(self.inputs[task_id]).is_file():
-
-                    with gw.open(self.outputs[self.inputs[task_id]]) as src:
-                        self.execute(task_id, task, src, **kwargs)
-
-                if isinstance(self.inputs[task_id], str) and Path(self.inputs[task_id]).is_file():
-
-                    with gw.open(self.inputs[task_id], **self.open_args) as src:
-                        self.execute(task_id, task, src, **kwargs)
-
-                else:
+                if isinstance(self.inputs[task_id], tuple) or isinstance(self.inputs[task_id], list):
 
                     with ExitStack() as stack:
 
                         # Open input files for the task
-                        src = [stack.enter_context(gw.open(fn, **self.open_args)) for fn in self.inputs[task_id]]
-                        self.execute(task_id, task, src, **kwargs)
+                        src = (stack.enter_context(gw_open(fn, **self.open_args)) if Path(fn).is_file()
+                               else task_results[fn] for fn in self.inputs[task_id])
+
+                        res = self.execute(task_id, task, src, task_results, attrs, **kwargs)
+
+                    # res = self.execute(task_id, task, self.inputs[task_id], task_results, attrs, **kwargs)
+
+                elif isinstance(self.inputs[task_id], str) and not Path(self.inputs[task_id]).is_file():
+                    res = self.execute(task_id, task, task_results[self.inputs[task_id]], task_results, attrs, **kwargs)
+                elif isinstance(self.inputs[task_id], str) and Path(self.inputs[task_id]).is_file():
+
+                    with gw_open(self.inputs[task_id], **self.open_args) as src:
+                        attrs = src.attrs.copy()
+                        res = self.execute(task_id, task, src, task_results, attrs, **kwargs)
+
+                task_results[task_id] = res
 
                 self._log_task(task_id)
 
-                if counter > 0:
-                    self._cleanup('task', self.tasks[counter-1][[0]])
+                # if counter > 0:
+                #     self._cleanup('task', self.tasks[counter-1][[0]])
 
                 counter += 1
 
-        for task_id, __ in self.tasks:
-            self._cleanup('pipeline', task_id)
+        # for task_id, __ in self.tasks:
+        #     self._cleanup('pipeline', task_id)
+
+        return res
 
 
-# class LandsatBRDFPipeline(GeoPipeline):
+# class IndicesStack(object):
 #
 #     """
-#     A pipeline class for Landsat BRDF
+#     A class for stacking spectral indices
 #
 #     Args:
 #         processes (tuple): The spectral indices to process.
@@ -473,140 +535,38 @@ class GeoTask(BaseGeoTask, GraphBuilder):
 #
 #     Example:
 #         >>> import geowombat as gw
-#         >>> from geowombat.core import pipeline
-#         >>> from geowombat.radiometry import RadTransforms
+#         >>> from geowombat.tasks import IndicesStack
 #         >>>
-#         >>> rt = RadTransforms()
-#         >>> meta = rt.get_landsat_coefficients('file.MTL')
+#         >>> task = IndicesStack(('avi', 'evi2', 'evi', 'nbr', 'ndvi', 'tasseled_cap'))
 #         >>>
-#         >>> task = pipeline.LandsatBRDFPipeline(('dn_to_sr', 'norm_brdf', 'bandpass'))
-#         >>>
-#         >>> with gw.open('image.tif') as src, \
-#         >>>     gw.open('sza.tif') as sza, \
-#         >>>         gw.open('saa.tif') as saa, \
-#         >>>             gw.open('vza.tif') as vza, \
-#         >>>                 gw.open('vaa.tif') as vaa:
-#         >>>
-#         >>>     res = task.submit(src, sza, saa, vza, vaa, sensor=meta.sensor, meta=meta)
+#         >>> with gw.open('image.tif') as src:
+#         >>>     res = task.submit(src, scale_factor=0.0001)
 #     """
 #
 #     def __init__(self, processes):
 #
-#         super().__init__(processes)
-#         self._validate_methods([rt, br, la])
+#         self.processes = processes
+#         self._validate_methods()
+#
+#     def _validate_methods(self):
+#
+#         args = [gw] * len(self.processes)
+#
+#         for object_, proc_ in zip(*args, self.processes):
+#
+#             if not hasattr(object_, proc_):
+#                 raise NameError(f'The {proc_} process is not supported.')
 #
 #     def submit(self, data, *args, **kwargs):
 #
-#         for i, func_ in enumerate(self.processes):
+#         attrs = data.attrs.copy()
+#         results = []
 #
-#             if func_ == 'dn_to_sr':
+#         for vi in self.processes:
 #
-#                 func = getattr(rt, func_)
+#             vi_func = getattr(gw, vi)
+#             results.append(vi_func(data, *args, **kwargs))
 #
-#                 if i == 0:
+#         results = xr.concat(results, dim='band').astype('float64')
 #
-#                     res = func(data,
-#                                *args,
-#                                sensor=kwargs['sensor'],
-#                                meta=kwargs['meta'])
-#
-#                 else:
-#
-#                     res = func(res,
-#                                *args,
-#                                sensor=kwargs['sensor'],
-#                                meta=kwargs['meta'])
-#
-#             elif func_ == 'norm_brdf':
-#
-#                 func = getattr(br, func_)
-#
-#                 if i == 0:
-#
-#                     res = func(data,
-#                                *args,
-#                                sensor=kwargs['sensor'],
-#                                wavelengths=data.band.values.tolist(),
-#                                out_range=10000.0,
-#                                nodata=65535)
-#
-#                 else:
-#
-#                     res = func(res,
-#                                *args,
-#                                sensor=kwargs['sensor'],
-#                                wavelengths=data.band.values.tolist(),
-#                                out_range=10000.0,
-#                                nodata=65535)
-#
-#             elif func_ == 'bandpass':
-#
-#                 func = getattr(la, func_)
-#
-#                 if i == 0:
-#
-#                     if kwargs['sensor'].lower() in ['l5', 'l7']:
-#
-#                         res = func(data,
-#                                    kwargs['sensor'].lower(),
-#                                    to='l8',
-#                                    scale_factor=0.0001)
-#
-#                     else:
-#
-#                         res = func(res,
-#                                    kwargs['sensor'].lower(),
-#                                    to='l8',
-#                                    scale_factor=0.0001)
-#
-#         return res
-
-
-class IndicesStack(object):
-
-    """
-    A class for stacking spectral indices
-
-    Args:
-        processes (tuple): The spectral indices to process.
-
-    Returns:
-        ``xarray.DataArray``
-
-    Example:
-        >>> import geowombat as gw
-        >>> from geowombat.tasks import IndicesStack
-        >>>
-        >>> task = IndicesStack(('avi', 'evi2', 'evi', 'nbr', 'ndvi', 'tasseled_cap'))
-        >>>
-        >>> with gw.open('image.tif') as src:
-        >>>     res = task.submit(src, scale_factor=0.0001)
-    """
-
-    def __init__(self, processes):
-
-        self.processes = processes
-        self._validate_methods()
-
-    def _validate_methods(self):
-
-        args = [gw] * len(self.processes)
-
-        for object_, proc_ in zip(*args, self.processes):
-
-            if not hasattr(object_, proc_):
-                raise NameError(f'The {proc_} process is not supported.')
-
-    def submit(self, data, *args, **kwargs):
-
-        attrs = data.attrs.copy()
-        results = []
-
-        for vi in self.processes:
-
-            vi_func = getattr(gw, vi)
-            results.append(vi_func(data, *args, **kwargs))
-
-        results = xr.concat(results, dim='band').astype('float64')
-
-        return results.assign_attrs(**attrs)
+#         return results.assign_attrs(**attrs)